feat: complete backend implementation with spring boot

+++ ecommerce_multivendor/src/main/java/com/ecom/Repository/DealRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DealRepo extends JpaRepository<DealEntity, Integer> {
+public interface DealRepo extends JpaRepository<DealEntity, Long> {
 }
